+++ plugins/CHANGELOG.md
@@ -13,6 +13,7 @@
 - **plugin-types**: Added `textBounds` property for text shapes
 - **plugin-types**: Added flag `throwValidationErrors` to enable exceptions on validation
 - **plugin-types**: Fix missing `webp` export format in `Export.type`
+- **plugin-types**: Added `fixedWhenScrolling` property for shapes
 
 ## 1.4.2 (2026-01-21)